拆分 settings_dialog.py: 提取 SecuritySettingsSection / ReminderPresetSection，1324 行减至 750 行
